requirements.txt

streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.19.0
google-generativeai>=0.4.1
